Refactor `xcmsRaw` to use generic `xcmsSource` methods.
`xcmsRaw` no longer has hard-coded loading of raw data; it uses children of the
`xcmsSource` class to provide the raw data it uses. This means that different
backends could be added freely simply by defining new implementations of the
`xcmsSource` method.

Currently, there are classes which handle reading from files to provide the same
level of functionality as the hard-coded version of `xcmsRaw`.